# src/backend/rest/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: FridoLean.io
host: localhost:3001
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
  /auth/signin:
    x-swagger-router-controller: auth
    post:
      tags:
        - auth
      description: Signs in a user
      operationId: signIn
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SignInBody'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/SignInResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /auth/signup:
    x-swagger-router-controller: auth
    post:
      tags:
        - auth
      description: Signs up a user
      operationId: signUp
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SignUpBody'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/SignInResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/users-kc/{keycloakID}':
    put:
      tags:
        - user
      description: Update a user’s information. It will return the updated user object.
      operationId: createKeycloakUser
      parameters:
        - name: keycloakID
          in: path
          description: keycloak ID of a user
          required: true
          type: string
          format: uuid
        - name: newData
          in: body
          description: Update User
          required: true
          schema:
            $ref: '#/definitions/UpdateUser'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserInformation'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/users/{userID}':
    x-swagger-router-controller: user
    get:
      tags:
        - user
      description: Get a user’s profile information. It will return the user object.
      operationId: getUser
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: userID
          in: path
          description: ID of a user
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserInformation'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - user
      description: Update a user’s information. It will return the updated user object.
      operationId: updateUser
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: userID
          in: path
          description: ID of a user
          required: true
          type: string
          format: uuid
        - name: newData
          in: body
          description: Update User
          required: true
          schema:
            $ref: '#/definitions/UpdateUser'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserInformation'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/NotFoundErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - user
      description: Delete a user’s profile. It will return the deleted user object.
      operationId: deleteUser
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: userID
          in: path
          description: ID of a user
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserInformation'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/users/{userID}/projects':
    x-swagger-router-controller: user
    get:
      tags:
        - user
        - project
      description: >-
        Get a user’s projects list. It will return the list of project id’s the
        user is associated with.
      operationId: getUserProjects
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: userID
          in: path
          description: ID of a user
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserProjects'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  UpdateUser:
    properties:
      password:
        type: string
      email:
        type: string
  UserInformation:
    properties:
      id:
        type: string
      email:
        type: string
  UserProjects:
    properties:
      projects:
        type: array
        items:
          $ref: '#/definitions/ProjectMeta'
...
  ProjectMeta:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
      visibility:
        type: string
        description: PUBLIC || PRIVATE
      lastEdited:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
      description:
        type: string
      members:
        type: array
        items:
          properties:
            userID:
              type: string
            email:
              type: string
            rights:
              type: string
...
  SignUpBody:
    required:
      - email
      - password
      - passwordRepeat
      - termsAndConditions
    properties:
      email:
        type: string
      password:
        type: string
      passwordRepeat:
        type: string
      termsAndConditions:
        type: boolean
  SignInBody:
    required:
      - email
      - password
    properties:
      email:
        type: string
      password:
        type: string
...
  NotFoundErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
        description: Describes which instance could not be found
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SignInResponse:
    required:
      - token
    properties:
      token:
        type: string
      success:
        type: boolean
      message:
        type: string
      user:
        $ref: '#/definitions/UserInformation'